merge instead of rebase
rebasing could require use of `--force` or other
workarounds when getting the licenses branch
updated from the user's work branch.
users might have rules in place on a repo that
prohibit force pushes.  merging up is safer and
easier to reason about, and gives users the option
on how to merge the changes into the work
branch - merge, squash, rebase