Eliminate nock tests
The updated actions SDK uses `fetch` and `nock` (used for mocking
network requests) only supports mocking fetch in beta. The beta
version of nock did not work right away.

The tests using nock were only testing a tiny bit of functionality,
so it's not worth trying to fix this at this time.